Add files via upload
Though this does seem to effectively map gff attributes to an exome chain file, it is very inefficient (~1 hr for the human exome). I am in the process of collating the chains differently in order to write each chromosome all at once and will upload it to the master branch once that is complete.